+++ CHANGELOG.md
@@ -26,6 +26,7 @@ You should also include the user name that made the change.
 	- e.g. `https://p1.a9z.dev/emoji/misskey.webp`
 	- remote: `https://p1.a9z.dev/emoji/syuilo_birth_present@mk.f72u.net.webp`
 - API: `user`および`note`エンティティに`emojis`プロパティが含まれなくなりました
+- API: `user`エンティティに`avatarColor`および`bannerColor`プロパティが含まれなくなりました
 
 ### Improvements
 - Push notification of Antenna note @tamaina